feat: Add kanji candidate support with IPADIC dictionary
- Integrate IPADIC dictionary for kanji conversion
- Refactor web app to support kanji candidates
- Add lattice-based path finding for kanji conversion
- Update data types for better memory usage (u32 for IDs, i16 for costs)
- Increase WASM memory pages for dictionary support
- Restructure React app with routing and new components

+++ src/core/WordEntry.zig
@@ -1,4 +1,4 @@
 word: []const u8,
-left_id: usize = 0,
-right_id: usize = 0,
-cost: isize = 0
+left_id: u32 = 0,
+right_id: u32 = 0,
+cost: i16 = 0
